Extract rumoca-input: InputEngine + SignalMapper + input config types
Moves the input engine, signal mapper, and all input/signal TOML config
types out of rumoca-sim-fb into a new rumoca-input crate. Matches the
naming-scheme proposal's per-axis family — this crate owns everything
rooted in the LocalStore / Path primitives, including the signal mapper
that consumes them.

- crates/rumoca-input: new crate
  - src/config.rs: LocalDef, DeriveSpec, InputConfig + all subtypes,
    SignalsConfig, SignalSpec (moved from sim-fb)
  - src/engine/: InputEngine + compile step (moved from sim-fb/input_engine/)
  - src/signal_mapper.rs: SignalMapper (moved from sim-fb/)
- rumoca-sim-fb: now depends on rumoca-input, drops gilrs direct dep.
  Config re-exports the moved types so existing call sites keep working.
- Test helpers in the moved files decoupled from SimFbConfig — each test
  parses a minimal local Subset struct instead.

31 rumoca-input tests pass (moved from sim-fb); workspace clippy clean.

Follow-up (phase 2b/c): split device-specific code out into
rumoca-input-gamepad and rumoca-input-keyboard using the InputDevice
trait (designed in phase 1, not yet implemented).
