+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
     "crates/rumoca-codec-flatbuffers",
     "crates/rumoca-transport-udp",
     "crates/rumoca-transport-websocket",
+    "crates/rumoca-input",
     "crates/rumoca-sim",
     "crates/rumoca-solver-diffsol",
     "crates/rumoca-solver-rk45",
@@ -71,6 +72,7 @@ rumoca-codec = { path = "crates/rumoca-codec" }
 rumoca-codec-flatbuffers = { path = "crates/rumoca-codec-flatbuffers" }
 rumoca-transport-udp = { path = "crates/rumoca-transport-udp" }
 rumoca-transport-websocket = { path = "crates/rumoca-transport-websocket" }
+rumoca-input = { path = "crates/rumoca-input" }
 rumoca-sim = { path = "crates/rumoca-sim" }
 rumoca-solver-diffsol = { path = "crates/rumoca-solver-diffsol" }
 rumoca-solver-rk45 = { path = "crates/rumoca-solver-rk45" }
