+++ crates/rumoca-input/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "rumoca-input"
+description = "Input engine + signal mapping for Rumoca lockstep simulations"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+rumoca-codec = { workspace = true }
+
+anyhow = { workspace = true }
+indexmap = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+toml = { workspace = true }
+
+# Temporarily held here; migrate to -gamepad / -keyboard device crates
+# once the InputDevice trait migration lands.
+crossterm = "0.28"
+gilrs = "0.11"
+
+[lints]
+workspace = true
