+++ crates/rumoca-sim-fb/src/lib.rs
@@ -5,8 +5,6 @@
 //! example path, viewer, and controller process wiring.
 
 pub mod config;
-pub mod input_engine;
-pub mod signal_mapper;
 pub mod sim_loop;
 
 use std::path::Path;
